pyproject.toml
[tool.poetry]
name = "root"
version = "0.1.0"
description = ""
authors = ["Khasanjon <khasanjon.eng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
pillow = "^10.0.1"
django-redis = "^5.3.0"
psycopg2-binary = "^2.9.7"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"
django-jazzmin = "^2.6.0"
django-cors-headers = "^4.3.0"
isort = "^5.12.0"

[tool.isort]
skip = "migrations"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
